More value types accepted for "xyz" attribute.
- Added more value types to the set of understood "xyz" values.
- Changed the graphicGraph.GraphPosLengthUtils accordingly.
- Added some methods to Vector2 and Vector3 to prepare the removal of
  ScalaGS.

+++ src/org/graphstream/ui/graphicGraph/GraphicNode.java
@@ -142,7 +142,7 @@ public class GraphicNode extends GraphicElement implements Node {
 			} else if (attribute.equals("z")) {
 				moveFromEvent(x, y, numberAttribute(newValue));
 			} else if (attribute.equals("xy") || attribute.equals("xyz")) {
-				float pos[] = nodePosition(this);
+				double pos[] = nodePosition(this);
 
 				moveFromEvent(pos[0], pos[1], pos[2]);
 			}
